requirements.txt
fastapi-htmx
python-dotenv
uvicorn
databricks-sql-connector[sqlalchemy]
sqlalchemy-databricks